+++ source/brailleDisplayDrivers/brailliantB.py
@@ -351,6 +351,7 @@ class BrailleDisplayDriver(braille.BrailleDisplayDriver):
 				"braille_previousLine": ("br(brailliantB):up",),
 				"braille_nextLine": ("br(brailliantB):down",),
 				"braille_routeTo": ("br(brailliantB):routing",),
+				"braille_selectRange": ("br(brailliantB):multiRouting",),
 				"braille_toggleTether": ("br(brailliantB):up+down",),
 				"kb:upArrow": ("br(brailliantB):space+dot1", "br(brailliantB):stickUp"),
 				"kb:downArrow": ("br(brailliantB):space+dot4", "br(brailliantB):stickDown"),
@@ -389,6 +390,7 @@ class InputGesture(braille.BrailleDisplayGesture, brailleInput.BrailleInputGestu
 
 		self.keyNames = names = []
 		isBrailleInput = True
+		routingIndexes: list[int] = []
 		for key in self.keyCodes:
 			if isBrailleInput:
 				if DOT1_KEY <= key <= DOT8_KEY:
@@ -401,12 +403,15 @@ class InputGesture(braille.BrailleDisplayGesture, brailleInput.BrailleInputGestu
 					self.dots = 0
 					self.space = False
 			if key >= FIRST_ROUTING_KEY:
-				names.append("routing")
-				self.routingIndex = key - FIRST_ROUTING_KEY
+				routingIndexes.append(key - FIRST_ROUTING_KEY)
 			else:
 				try:
 					names.append(KEY_NAMES[key])
 				except KeyError:
 					log.debugWarning("Unknown key with id %d" % key)
+		if routingIndexes:
+			routingIndexes.sort()
+			self.cellIndexes = routingIndexes
+			names.append(self.idForCellCount(len(routingIndexes)))
 
 		self.id = "+".join(names)