+++ source/brailleDisplayDrivers/nlseReaderZoomax.py
@@ -234,6 +234,7 @@ class BrailleDisplayDriver(braille.BrailleDisplayDriver):
 				"braille_previousLine": ("br(nlseReaderZoomax):d1",),
 				"braille_nextLine": ("br(nlseReaderZoomax):d3",),
 				"braille_routeTo": ("br(nlseReaderZoomax):routing",),
+				"braille_selectRange": ("br(nlseReaderZoomax):multiRouting",),
 				"kb:upArrow": ("br(nlseReaderZoomax):up",),
 				"kb:downArrow": ("br(nlseReaderZoomax):down",),
 				"kb:leftArrow": ("br(nlseReaderZoomax):left",),
@@ -264,11 +265,11 @@ class InputGesture(braille.BrailleDisplayGesture, brailleInput.BrailleInputGestu
 				self.dots = groupKeysDown >> 8
 				self.space = groupKeysDown & SPACEBAR_KEYS_MASK
 			if group == DeviceCommand.ROUTING_KEYS:
-				for index in range(braille.handler.display.numCells):
-					if groupKeysDown & (1 << index):
-						self.routingIndex = index
-						names.append("routing")
-						break
+				self.cellIndexes = [
+					index for index in range(braille.handler.display.numCells) if groupKeysDown & (1 << index)
+				]
+				if self.cellIndexes:
+					names.append(self.idForCellCount(len(self.cellIndexes)))
 			else:
 				for index, name in enumerate(COMMAND_RESPONSE_INFO.get(group).keys):
 					if groupKeysDown & (1 << index):